refactor: 테스트 메서드 클래스 이동
